SDR-241 refactor: 패키지 및 parameterized test를 사용하도록 변경
